Small rearranging of IBMQJob class
